fix: the money path was float, not Decimal — ledger drifted 2 paisa
CLAUDE.md rule 3 says money uses DECIMAL, never float. The columns always
did: Numeric(12,2) throughout. The service layer did not — it read every
balance out with float(), did the arithmetic in binary floating point, and
wrote the result back. `account.cash = round(float(account.cash) - cost_basis, 2)`
is the whole defect in one line.

MEASURED, not asserted. Running the pre-fix module against 1,293 round trips
of randomised NSE-shaped fills (prices 50-5,000, seed 7) and comparing it to
the identical arithmetic in Decimal:

    ledger cash (pre-fix) : 289697.84
    exact Decimal cash    : 289697.82
    divergence            : 0.02   — and realized P&L, by the same 2 paisa

Same run against this commit: 0.00 on both.

Two things I want on the record rather than buried, because the size of a
defect matters as much as its existence:

  * It is small and slow. It needs the round-to-paisa boundary walked from
    both sides, which randomised prices do and a repeated fixed price pair
    does not — a fixed-price probe over 2,000 round trips diverged by 0.00.
    This was never going to show as a visibly wrong number on the page. It
    showed as a ledger that quietly stopped being reconcilable.
  * paper_trading.py's docstring claims the P&L identity "holds exactly, at
    every point". On the old code its residual was ~1e-10, not a rupee
    figure. So the identity claim was very nearly true; the cash and
    realized-P&L divergence above is the stronger finding, and I had earlier
    described the identity residual as the headline. It was not.

Also fixed, found while converting:

  _is_rebalance_day() counted ANY prior AITradingRun row this month as the
  month's rebalance, including a failed one. run_date is UNIQUE and the
  scheduler writes the row before trading, marking it "failed" if the cycle
  raises — so a single transient failure on the 1st (a yfinance timeout, a DB
  blip) left the account running stop/target checks only for the rest of the
  month: no entries, no regime trim, and nothing anywhere saying so. Now only
  a "complete" run consumes it.

  _sell_pass() called _mark_prices() once per open position inside the loop.
  Beyond the N+1, it meant two holdings in the same cycle could be judged
  against ticks captured seconds apart — one exit decision made on a tape the
  next one no longer saw. Marks are now taken once for the pass.

  Mapped[float] on Numeric columns in models/paper_trading.py was a type lie
  — SQLAlchemy returns Decimal regardless — and it is the lie the service
  layer believed. Now Mapped[Decimal].

  backend/test_momentum.py opened SessionLocal() at import time and pytest
  collects test_*.py by filename, so a bare `pytest` at the backend root ran
  the suite against the PRODUCTION database. Moved to
  scripts/inspect_momentum.py; pytest.ini pins collection to tests/.

Tests: 25, up from 2. test_randomised_round_trips_match_an_exact_shadow_ledger
is the one that discriminates — verified failing on the pre-fix module with
Decimal('289697.84') == Decimal('289697.82'), passing here. The two cadence
tests were likewise verified failing without the status filter.

pytest is in a new requirements-dev.txt, not requirements.txt: the Dockerfile
installs the latter into the production image and a test runner has no
business shipping to the box that trades. The runtime stack is unchanged.

+++ backend/app/core/scheduler.py
@@ -33,6 +33,7 @@ import time
 from contextlib import asynccontextmanager
 from concurrent.futures import ThreadPoolExecutor
 from datetime import date as date_type, datetime, timedelta
+from decimal import Decimal
 
 import yfinance as yf
 from sqlalchemy import func
@@ -1036,7 +1037,7 @@ def _ai_trading_done_today(as_of: date_type) -> bool:
         db.close()
 
 
-def _ai_trading_notification_text(result, equity: float | None) -> str:
+def _ai_trading_notification_text(result, equity: Decimal | None) -> str:
     kind = "monthly rebalance" if result.rebalanced else "daily stop/target check"
     lines = [f"<b>AI Trading — {result.as_of.isoformat()}</b> ({kind}, regime: {result.regime})"]
     for b in result.bought:
